multiple evm-uniswap routes
