+++ .env.example
@@ -7,51 +7,65 @@ PORT=5000
 # ipv6 format for locahost ["::ffff:127.0.0.1", "::ffff:1", "fe80::1", "::1"]
 IP_WHITELIST=
 
-HUMMINGBOT_INSTANCE_ID={client_id}
-
-# Celo
-
-# Terra
-# - mainnet: https://lcd.terra.dev
-# - mainnet chain: columbus-4
-# - testnet: https://tequila-lcd.terra.dev
-# - testnet chain: tequila-0004
-TERRA_LCD_URL={testnet_lcd_url}
-TERRA_CHAIN={testnet_chain_id}
+HUMMINGBOT_INSTANCE_ID=ae814aa806fff4de741792d15d4684aaa8009580
 
 # Ethereum
-# - chain: mainnet, kovan, etc
+# - chain: mainnet, kovan, xdai, etc
 # - rpc url: infura or other rpc url
 # - token list: erc20 token list source (ref: https://tokenlists.org/)
-ETHEREUM_CHAIN={chain}
-ETHEREUM_RPC_URL=https://{chain}.infura.io/v3/{api_key}
-ETHEREUM_TOKEN_LIST_URL=https://wispy-bird-88a7.uniswap.workers.dev/?url=http://tokens.1inch.eth.link
+ETHEREUM_TOKEN_LIST_URL=https://tokenlists.org/token-list?url=https://raw.githubusercontent.com/compound-finance/token-list/master/compound.tokenlist.json
+ETHEREUM_RPC_URL=https://mainnet.infura.io/v3/64f5e058d11f4bc686b31ab4d5430688
+ETHEREUM_CHAIN=mainnet
+# Ethereum: EthGasStation
+# API key for defipulse.com gas station API
+# Gas level you want to use for Ethereum transactions (fast, fastest, safeLow, average)
+ENABLE_ETH_GAS_STATION=false
+#ETH_GAS_STATION_API_KEY={apikey}
+#ETH_GAS_STATION_GAS_LEVEL=fast
+#ETH_GAS_STATION_REFRESH_TIME=60
+MANUAL_GAS_PRICE=1
 
-# Balancer
-# subgraph_chain
-# Reference: https://docs.balancer.finance/sor/development#subgraph
-# - mainnet: balancer
-# - kovan: balancer-kovan
-# Note: REACT_APP_SUBGRAPH_URL used by @balancer-labs/sor
-REACT_APP_SUBGRAPH_URL=https://api.thegraph.com/subgraphs/name/balancer-labs/{subgraph_chain}
-
-# exchange_proxy:
-# Reference: https://docs.balancer.finance/smart-contracts/addresses
-# - mainnet: 0x3E66B66Fd1d0b02fDa6C811Da9E0547970DB2f21
-# - kovan: 0x4e67bf5bD28Dd4b570FBAFe11D0633eCbA2754Ec
-EXCHANGE_PROXY={exchange_proxy}
-
-# Uniswap
+# Ethereum Uniswap
 # Reference: https://uniswap.org/docs/v2/smart-contracts/router02/
 # UniswapV2Router02 is deployed at 0x7a250d5630B4cF539739dF2C5dAcb4c659F2488D on the Ethereum mainnet, and the Ropsten, Rinkeby, Görli, and Kovan testnets. 
 # It was built from commit 6961711.
 UNISWAP_ROUTER=0x7a250d5630B4cF539739dF2C5dAcb4c659F2488D
 # allowed slippage for swap transactions
-UNISWAP_ALLOWED_SLIPPAGE=1.5
+UNISWAP_ALLOWED_SLIPPAGE=0.5
 # restrict updating pairs that have no reserves or failed for 5 minutes
 UNISWAP_NO_RESERVE_CHECK_INTERVAL=300000
 # cache info about pair for 1 second
-UNISWAP_PAIRS_CACHE_TIME=1000
+UNISWAP_PAIRS_CACHE_TIME=5000
+
+# EVM
+# - chain: mainnet, kovan, xdai, etc
+# - rpc url: infura or other rpc url
+# - token list: erc20 token list source (ref: https://tokenlists.org/)
+EVM_CHAIN=bsc
+EVM_CHAIN_ID=56
+EVM_FEE_TOKEN=BNB
+EVM_ENS_RESOLVER=0x0000f22ffe0866ffb8834600dad9259cf4956853
+EVM_RPC_URL=https://bsc-dataseed.binance.org/
+EVM_TOKEN_LIST_URL=https://almostfree.sfo2.digitaloceanspaces.com/bsc.json
+EVM_UNISWAP_GAS_LIMIT=1237500
+EVM_MANUAL_GAS_PRICE=5
+
+# EVM_UNSWAP
+# Arbitrary Uni-v2 DEX
+# ***MUST BE ABI COMPATIBLE w/ UNISWAP-v2***
+# Reference: https://uniswap.org/docs/v2/smart-contracts/router02/
+# shared
+EVM_UNISWAP_ALLOWED_SLIPPAGE=0.5
+EVM_UNISWAP_NO_RESERVE_CHECK_INTERVAL=300000
+EVM_UNISWAP_PAIRS_CACHE_TIME=5000
+# Pandaswap
+EVM_UNISWAP_0_ROUTER=0x29D1Adbb65d93a5710cafe2EF0E8131f64E6AB22
+EVM_UNISWAP_0_FACTORY=0x9Ad32bf5DaFe152Cbe027398219611DB4E8753B3
+EVM_UNISWAP_0_INIT_CODE_HASH=0xb785cebe4aa62be64e9e412e79c479c58578c1c3d0ffbb4602a9217d8b289559
+# PancakeSwap v2
+EVM_UNISWAP_1_ROUTER=0x10ED43C718714eb63d5aA57B78B54704E256024E
+EVM_UNISWAP_1_FACTORY=0xcA143Ce32Fe78f1f7019d7d551a6402fC5350c73
+EVM_UNISWAP_1_INIT_CODE_HASH=0x00fb7f630766e6a796048ea87d01acd3068e8ff67d078148a3fa3f4a84f69bd5
 
 # cert
 CERT_PATH={full_path_to_certs_folder}
@@ -64,20 +78,3 @@ LOG_PATH=/Users/hbot/hummingbot_files/hummingbot_logs
 # GMT offset for logging (alpine docker image default to UTC timezone)
 # -0800, -0500, +0200, +0800
 GMT_OFFSET=+0800
-
-# EthGasStation
-# API key for defipulse.com gas station API
-# Gas level you want to use for Ethereum transactions (fast, fastest, safeLow, average)
-ENABLE_ETH_GAS_STATION=true
-ETH_GAS_STATION_API_KEY={apikey}
-ETH_GAS_STATION_GAS_LEVEL=fast
-ETH_GAS_STATION_REFRESH_TIME=60
-MANUAL_GAS_PRICE=100
-
-# Balancer Config
-BALANCER_MAX_SWAPS=4
-
-
-# Perpetual Finance Provider URL
-# default:  https://dai.poa.network , https://rpc.xdaichain.com, etc
-XDAI_PROVIDER={providerUrl}
